+++ src/com/jayantkrish/jklol/training/LogFunction.java
@@ -12,17 +12,17 @@ public interface LogFunction {
 
   public void log(Assignment example, FactorGraph graph);
 
-  public void log(int iteration, int exampleNum, Assignment example, FactorGraph graph);
+  public void log(long iteration, int exampleNum, Assignment example, FactorGraph graph);
   
   public void logMessage(Object message); 
   
-  public void logParameters(int iteration, SufficientStatistics parameters);
+  public void logParameters(long iteration, SufficientStatistics parameters);
 
-  public void logStatistic(int iteration, String statisticName, double value);
+  public void logStatistic(long iteration, String statisticName, double value);
 
-  public void notifyIterationStart(int iteration);
+  public void notifyIterationStart(long iteration);
 
-  public void notifyIterationEnd(int iteration);
+  public void notifyIterationEnd(long iteration);
 
   public void startTimer(String timerName);
